% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_squares.R
\name{add_squares}
\alias{add_squares}
\title{Title}
\usage{
add_squares(num1, num2)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x}^2 and \code{y}^2
}
\description{
Description
}
\details{
Details
}
\examples{

add_squares(1, 2)

add_squares(1, -2)

}
\author{
Kurt Wirth
}
