% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_sq.R
\name{my_sq}
\alias{my_sq}
\title{Title}
\usage{
my_sq(num3)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x}^2 and \code{y}^2
}
\description{
Description
}
\details{
Details
}
\examples{
add_squares(1, 2)

add_squares(1, -2)

}
\author{
Kurt Wirth
}
\keyword{internal}
